+++ inc/com/centreon/logging/syslogger.hh
@@ -42,10 +42,10 @@ namespace              logging {
                        ~syslogger() throw ();
     void               close() throw ();
     void               log(
-                         unsigned long long types,
-                         unsigned int verbose,
+                         uint64_t types,
+                         uint32_t verbose,
                          char const* msg,
-                         unsigned int size) throw ();
+                         uint32_t size) throw ();
     void               open();
     void               reopen();
