+++ test/exceptions/basic_insert_ulonglong.cc
@@ -24,14 +24,14 @@
 using namespace com::centreon::exceptions;
 
 /**
- *  Check the basic insert unsigned long long.
+ *  Check the basic insert uint64_t.
  *
  *  @return 0 on success.
  */
 int main() {
   basic ex;
-  ex << static_cast<unsigned long long>(0);
-  ex << static_cast<unsigned long long>(ULLONG_MAX);
+  ex << static_cast<uint64_t>(0);
+  ex << static_cast<uint64_t>(ULLONG_MAX);
 
   std::ostringstream oss;
   oss << 0 << ULLONG_MAX;
